requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-openai==0.2.8
langgraph==0.2.33
pydantic==2.9.2
python-dotenv==1.0.1
django==5.1.2        
